Preparing to split modules to packages
The `models`, `tables`, 'forms', and `views` modules are rather unwieldy
and will soon grow even further.  The django app generator template creates
these as module files, but it does not prohibit them from becoming
packages rooted at the same names.

The objective is to split the exisiting monolothic files into smaller
blocks, and assemble them via __init__.py imports such that pre-existing
import statements elsewhere in the code base remain unchanged.

New classes and functions created during this work will then get added
as additional package contents.  The hope is that this will make it
easier to reason abou the scope and degree of change, since it will
largely happen in new files rather than additions to the existing
history's refactored files.

Although this commit and the next may involve a fair number of moved
lines of code, the intent is not to functionally change anything.  Black
and the on save code cleanup may re-arrange imports, update formatting,
and alter whitespace, but functional changes are not to be attempted
here.

There are some constant declarations and utility methods in these files
that are shared, but have no major dependencies of their own.  Imports
across major component files for the sake of acquiring constants and
utility methods can lead to some problematic side effects in Python, so
this refactoring will relocate some of those methods and constants to
reports.utils and reports.values.  These should be the only changes to
consumer import depenencies observed in this commit and the next.