IO-258 Batch Upload
Share the same volume when delegating uploads through NGinx

Also allow for development without image rebuild by mapping the
development file over the image's fixed copy in
docker-compose.override.yml

First stabilization milestone

-- Drag and drop loads the view
-- LocalStorage remembes unsubmitted uploads
-- Need to wire to the async Upload form method and render status
message still
-- Styling is wonky--where's the bootstrap look and feel?

DataTables 1.10.24 Dependency

Yes, it is highly modular.

Also organized first by feature then by CSS/JS, so it doesn't fit so
well in the way that /static is currently organized, so I created a
/static/vendor tree for such things.

Actually, this may be a more sensible way of organizing third party
dependencies, leaving /static/js and /static/css for our own artifacts
so its easier to discern what we're importing an what we wrote
ourselves.

Non-milestone commit

Checkpointing because this has some positive work, but it is troubled by
an order of event trigger versus table population.

REalized I have an alrerntative that will smooth things out, but I want
to freeze a point of reference in case there are comokication

Speicific problem is that rows are publishing events that affect their
visibility rendering before the changes that cause them are present in
the table.

One mistake is that I have been trying to use a no-op edit to cause the
server to pull AJAX changes after they have been made.  IF those changes
involve a styling change (e.g. read/write to read/only, or
disabling/enabling buttons), then an event is required to inform the
table that there is somethin differnet to accomodate, but that event
is currently falling before the change propogation has reached
the same lcoation.

I found I could also tell the esrver to pull changes without an
edit operation.  In doing so, I can register a callback to be
informed when the data has arrived, and that is a much better place
to address post-change stlying than a handler that has to run.

The other prolem is that the Edit API can only be used to manipulate
row-related state that is equally available for hte end user to edit.
Backend server calls can send additional row-associated data that the
Table is configured to respond to, but the editor disregards entirely.

This ability is needed here in orer to carry the state machine from
"drop zone" -> "upload queue" -> "upload completed".  Right now, that
state is implemented as part of the editable model.  I can mark it
read-only, but even then it shows uo in detai l forms and that is
bound to be confusing for en users.

In this case, our "server" is not remote--it is an in-client
object bound to localStorage.  Reloading the whole table to affect only
one or two rows is not ideal, but when a transition targets N rows,
such as the "launch batches" button moving N copleted uplaod forms
to the work queue, it is best to be able to update the N rows in
the backend and then issue a single table refresh rather than to
refresh the table one row at a time, because that's the difference
between O(N) and O(N^2) time complexity!

Stablizing!

The Javascript client side seems to be mostly on target.  I am now
gettign a server-side error.  Checkpointing because I don't want to risk
losign this place!

Well well well, it works...

Minimally so at least...  Files reach the other side and get imported,
but the UI's still not so great at delivering feedback about their
progression.

Some times rows that are in the cache don't come back on return, and
sometimes the state notifications seem to get dropped.  Ther was earlier
a problem where the Begin button would not get enabled when you finished
filling out a form but I think that may be resolved...

These look like manifestations of race condtions when interacting with
multiple files, a side effect of having simplified the eventing to work
in terms of full page refreshes is that if the library has some measure
of asynchronous processing, it could be that two or more paint requests
arrive at the same time, only the last of which reflects the true
complete state.  All start rendering at about the same time, but without
something to guarantee that the last paint to start is also the last
paint to finish, its conceivable that you might reach the end of such a
series of events at an intermediate version of truth...

IO-428 Corrected remaining known bugs

Various bug fixes

Corrected issue where ast known good field would reset on any other
field edit.

Corrected issue where the inline edit field would not initially show the
previous value on edit.

Prevented incorrect determination that an upload was successful when it
reutrns a 200 OK but the message describes the reason an error occurred.

Enabling nginx accelerartion

Async upload operations for nginz

Hashes for conflict detection

Final bug fixing correcting issues that manifested post-rebase