IO-430 Regression Corrections
Fixing some regressions related to this ticket discovered while drafting
tests:

-- TASER linking had broken and was briefly linking to the report detail
page like all other links instead of the related Jira ticket.

-- Could not upload non-Genexus device archives because the conditional
logic designed to prevent us from searching for Genexus lane information
was only preventing the search for information to link to a
ValkyrieArchive extension, but was not preventing the creation of that
extension.  With nothing filled in, the extension violated NOT NULL
constraints and caused the upload to abort.
-- The routine now returns quickly rather than attempting to jump
over logic since there is no further logic in that handler once a
non-Genexus device is recognized as such.

-- Could not save edits to Genexus devices that had not yet added a
ValkyrieArchive extension.  The handler had been assuniung that the only
edits to a Genexus device it would always follow their initial creation
and decoration.  It therefore was presunuing that not being a new
creation implied the ValkyrieExtension could be used without first
verifying its existance.
-- Corrected this bug by adding the conditional logic and allowing
for lazy registration if we have not yet migrated a given Archive before
a user next saves editting chagnes.

-- Filtering was susceptible to crashing if a filter was asserted before
Django had opportunity ti cache an Archive's ValkyrieArchive extension.
Corrected this by ensuring that we have attached a `select_related` call
to the Query Set and ensured that hydrating an Archive will always also
hydrate its ValkyrieExtension.
-- ValkyrieArchive uses the master Archive record's primary key as
its foreign key to make this a more efficient join--both tables can be
accessed concurrently instead of in serial when perfoming a
lookup-by-id.

Also added some sorely needed logging for when save requests fail to the
clientside app.  Missing this made a moire challenging debug opertion.