IO-437 Enable functional Bootstrsp 3 tsbs for samole names with periods
HTML's constraints on legal id attribute values permits ids that match
this regular expression:

```
	/^[a-zA-Z][-\w:.]*$/
```

An unfortunate reality here is that HTML supports use of identifiers
that include periods anywhere except their first character.  Its
unfortunate becsuse period is also used as a separator between element
class names, making any element IDs that use a period in their names
inherrently ambiguous.

Consider the selector for am HTML element with id="this.that".  A
selector by id for this element would be given as "#this.that".  Now
consider an element with id="this" and class="that".  Its selector is
also $("#this.that").  In terms of which of the specified nodes JQuery
would target, the id & class pair based interpreration has won out and
is the interpretation chosen to resove the ambiguity.

This bit of background is the reason that Bootstrap Tabs do not work,
even when otherwise configured correctly, fail to trigger the loading of
depdenent panel views if those panel's idenifiers bear one or more
periods.  And because Inspector's QC Status Check diagnostics derives
the internal ID used to link each sample's data table to a navtab
bearing its display name, that menas that the navigational tabs for
experiments that use a period their names cannot use the resulting
navbar to toggle between its' sample-specific data tables.

This interface would also have been broken for any experiment that
used charcters in its name that are not valid in an HTML id, such as
"@", "#", ",". and many others.

The inspector Django app has no previoud template tasgs of its own,
so this change introduces a new template tag library.  Its first member
is a filter that rewrited an input string such that its output is a
vslid HTML Id that does not use periods, even if its input did.  It
consists of two logical rules:

1)  Replace any '.' or illegsl HTML id character with an underscore.
2)  If the result of (1) does not have begin with an allowed character
    for HTML identifiers, prepend the identifier with an 'x'.

The convention used here is simple by design because name mangling to
derive legal names may need to happen in other places and still be
consistent.  Any rule thst reduces the size of a countaby infinite
domain of langusge-compliant sdtrings to a smaller countsbly infinite
range of strings will inevitsbly have name collisions such that are not
unique even if the source strings were.

Apache Kafka has a similar problem where '_' and '.' in topic names are
treated as identical charascters "for historical reasons".  Mapping one
of those characters to a highly unlikely substring might sharply cut the
likelihood of a customer actually needing to use a pari f  names that
conflict but it could never eliminate the possibility, and that
complexity would have actually made it harder to explain to customers
what meaasure they could use to protect themselves from ambiguous names.

In our case, asking cutomers to not use periods at all would probably
be perceived as too much to ask.  The name manging done here now allows
a simpler request to be made.  Customers may use both '.' and '_' in
experiment naming, but must be careful that use does not create a
situation where two or more unique names differ only by one or more
transpositions between '.' and '_'.

An example of a problematic set of names can help illustrate.  Imagine
an experimetnm with two numeric variables, one with 1.0 or 1.2 units and
the other with 2.5 or 5.0 units.  The caresian product has four pairs of
these two values.  By ignoring decimals on the two unit-divisible values
(1.0 and 5.0), this could lead to the four following names:

    Exp-1_2.5, Exp-1_5, Exp-1.2_2.5, and Exp-1.2_5

The filter created here would derive the name "Exp-1_2_5' for two
different inputs.  Fortunately, accomodating the possbility can be done
in many easy ways, incuding but by no means limitted to

-- Use two underscores to separate the two variable values, yielding
   Exp-1_2__5 and Exp-1__2_5 for the prevously clashing names.
-- Use a hyphen instead of an underscore to separate the vales:
   Exp-1_2-5 and Exp-1-2_5 now result.
   -- Note that using some other separators would not work, as
      many of those also get maapped to '_'!