This commit escaped the rc-8 release, but is recommended.
Asynchhronous I/O in nginx creates a thread pool for asynchronous
reading, but that is not going to be of much value for bulk upload,
and I have some observations that suggest this can cause a race
condition that leads to the bulk upload sending a request on to
the upstream without the file attachment headers in place.

It's more trouble than its worth to try using this yet, so I
recommend removing the 'aio thread' option from our bulk upload
directive blocks.