IO-258 More UX fixes from review
This adds:

-- Responsive adjustments don't wreck the UI with small screen sizes.
-- Implwmented a scrollable drop zone at smaller sizes rather than
having the interface collapse in on itself.
-- Clicking a Green report now loads in a new tab
-- Full filename is displayed with no more ellipsis truncation.
-- Corrected an issue that would sometimes gradually increase the number
of uploads in parallel.  This will keep it at one.
-- Implemented left-right scroll so that the table contents and header
stay inline with small screen sizes.
-- Probably something else.  If I recall I will add it as a comment.
-- Diagnostics working with tar format files could create an
inaccessible diagnostic results directory because it was attempting to
set permission bits in octal, but was its input was actualy in decial,
leading to a very weird write-only directory access. (no read, no
trverasal)

IO-258
IO-428