IO-383 Format Date Column in PST on Archive List View.
Render timestamp in Pacific time zone.  This can be adapted to account
for other time zones, but we need to settle on the interface aspects of
discovering what timezone an end user is in to account for that
variability.

Reverting to use UTC is not difficult and may be prefferable to hard
coding PTS as it has a deeper repoutation for being a reference time
zone.