IO-441 Handle exceptions for errors that augment but do not break ChipStatus test
When this file is missing, I'm also seeing BaseCaller.json from the
BaseCallerActor is also missing and triggers another exception
downstream.

From what I can tell, if the report is run by a Genexus device, it's
message on the report summary page contains "Final Reads {}" based on
data found on counts found in `BaseCaller.json` and the report content
contains "Total Reads {}" based on the data found on counts in
`ionstats_basecaller.json`.

For other instrument types, the "Total Reads {}" message is used both on
the summary reporting view and inside the detailed diagnostic report
itself.  In both cases, the `total_reads` value that gets formatted into
this string message also gets passed to the detailed report template,
but it is not referenced there so it seems the error handling can get
away with:

```
    total_reads = None
    total_reads_message = "Total Reads Unknown"
```

Between these two code regions there are another three other variables
that are assigned, but never used, and one that is only used to provide
the formatted number for the Genexus-specific "Final Reads {}" message.
Since the error handling case is using 'unknown' in lieu of a formatted
number taken from `base_caller_read_stats['valid'], I think it is fair
for the error handling to populate these unused references with `None`
since these are each storing a measurement and there is, in fact,
nothing to measure.