IO-401: Allow customer to use underscore when naming Chef devices
Current extraction of date string from filename makes an assumption
about the number of underscore characters betweEn the start of the
filename and the start of its timestamp section.

The problem with this strtategy is that if the customer can introduce
underscores in the region of the larger string where underscores are
being precisely counted, that will throw off the ability to make
matches.

This change capitalizes on the fact that all customization a customer
can perform takes effect _before_ the date.  Instead of tokenizing
and counting tokens, this approach uses a regular expression
matching the valid date format and extracts the last such match to
that pattern.