IO-401 Reuse working solution from inspector_utils instead
The previous commit for this issue does contain a working solution, but
while implementing test cases covering it I came across another
pre-existing solution to the same problem in the common utilities that
is used by two other existing Diagnostics, Kit_Details and Run_Details,
both also Chef tests.

That routine is similar to the code that was initially replaced here.
It is avoding problems with the inability to guarantee how many _'s will
occur between the far left of the whole path name and the data field,
but in a different way.  Instead of a regular expression matching at the
end of the string, it is using rsplit limitted to the predictable number
of _'s starting from the far right.  This works for the same reason the
regular expression approach did--ther is no unpredictable content
between the date and the far right end of the string being parsed.

This commit replaces my new solution with a call to the existing helper
method from common insector utils that works and also adds the content
of this cases' path value to the test case where that common utility
method is validated.

Incidentally, it was possible to predict that this method would work
solve this case by observing the successfully parsed date string in the
Chef_Run_Details diagnostic for the input reported in Jira's IO-401
ticket.