IO-451: Improve maintainability and account for pre-diagnostic re-unpacking
The nested if/else blocks after locating any nested archive or report
PDF file was originall written with leaf nodes for all four possible
permutations of present/not present.  But the implementataion for the
present cases was the same regardless of whether the other condition was
present or not, so this just caused an arbitrary duplication of code
that could become a problem the first time it needs to change--there is
too much risk that someone might update one manifestation but not the
other.

This commit simplifies that handling to happen in serial rather than
nesting the conditional checks.

The previous code also implemented some sanity checking by refusing to
link a report PDF or unpack a nested archive if it found more than one
such candidate, but it did so by returning early, preventing a single
occurence of the other special case from being acted on.  An ambiguous
report PDF should not, for example, prevent the implementation from
unpacking a nested archive.  Instead of returning early, this change now
utilized an ambiguity detected marker value that defers action and
allows a non-ambiguous occurence of the other kind to take effect.

Finally, the earlier implementation was incapable of repeating the
archive extraction was attemtped at the start of every diagnostic
re-run.  The conditions created by the first unpacking leaves the
directory looking like the single-packed archive scenario that disables
this special case handling.  To make it possible to distinguish between
the first pass for a single-packed archive and the second or later pass
for a double-packed archive, the first pass for a double-packed archive
now leaves a symlink at a well-known filename.  After bypassing the
search because there are enough files to account for a single-packed
archive, this code now tests for the existance of that symlink and uses
it to re-do the double-packed archive unpacking at the same time that
any single-packed archive would have been unpacked again.