IO-451: Satisfy some SonarLint warning markers
SonarLint dislikes the cyclomatic complexity of the extract_archive
method and an easy way to reduce that is to isolate the nested try
blocks allowing zip archives to have .tar name suffix and vice versa to
distinct methods.

This has the added advantage of being reusable to similarly consider the
possibility that a nested archive could be created with the wrong file
extension for its actual format without repeating the large block of
nested exception handling again.

SonarLint also dislikes throwing un-subtyped Exception instances, so
I've replaced one such occurence in extract_archive with a more suitable
ArchiveWorkspaceError type.

Isolating this from the previous commit because it is more stylistic in
nature whereas the previous commit is focused on less subjective
necessity.