IO-449: Tolerate possibility of missing Bead Density file
Genexus archives are meant to have a bead density file, but in the off
chance that we should receive one that is missing this file, that should
not be a sufficient cause to fail the artifact import.

The previously reviewed submission for IO-449 did not tolerate the
possibility that no BEAD_DENSITY file might be found and would have
caused an OSError to occur by symlinking to None if that occurred.  That
exception was uncaught and would have failed the entire archive import
step.

This corrects that by testing the possibility that searching for a
BEAD_DENSITY_FILE_NAME might return None, and instead of attempting a
doomed symlink call, the handler now just logs the observation and
returns.  This will later cause broken image links on the instrument
tracker since the symlink its expecting to serve an image on the list
view with will not actually exist, but this would seem to be expected
behavior since an expected image file genuinely did not exist.