IO-440: Render Report PDF in place of Loading Density/Summary/Samples
There is no easy way to split the "Diagnostic" component from the
"Summary" subunit because they are rendered together by client-side Vue.

However, these server-side template can simply excxlude the Summary
content when it knows from the context that it is going to paint the PDF
viewer immediately after.

IO-451 has already arranged for the Genexus run report PDF to be
symlinked at /report.pdf under the archive root, which is where the PDF
viewer for S5 is already expecting to conditionally render if such a
file exists.

This completes the IO-440 request to have that PDF repalce the Loading
Density/Samples/Summary tables when present, and does do simply by
excluding those content elements on the same condition causing the PDF
viewer to appear.