IO-250: Tolerate MacOS extraction followed by recompression of CSA
When a Mac is used to decompress and then repackage a CSA, the file ends
up with its contents nested inside an inner directory.  The resolution
here is to detect the naming convention for that directory, which is
generally a strict prefix of the actual archive filename, and move its
contents up to the archive_root directory as part of the unzipping
process.