IO-441: Paint calibration.html, pix_offset_spatial, and pix_offset_histogram
The guard expressing whether chip_images.html was or was not found in the
CSA was previously used to control not only wheterh its own <iframe> was
included, but also whether the hyerlink to calibration.html and the image
tags painting pix_offset_spatial.png and pix_offset_histogram.png appear in
rendered output or not.

All four of these are uniquely occurring or non-ocurring files, so I narrowed
the scope of a guard testing for chip_images.html to only span inclusion
of its <iframe> tag, and then created three new context properties and
used each of them to make some portion of content unique to its file
condotionally included.  The entire <div> segment these three additional
items appear in is now conditional on at least one of the three member
items having been present.