IO-445/IO-457: Prune overwritten data points from flow sequence logs
Ok, so IO-457 is indeed related to IO-445, and the "fix" I'd attempted
for IO-445 was something the actual problem had me fooled into thinking
was solving something, but it turns out I was just masking one symptom
of the real underlying cause.

There reason that the flow logs are failing to parse in this diagnostic
is not that there are a few random entries that have something slightly
different as I thought.  In the issue reported, there were three data
point that were returned by the explog parse a little wierd: acq_0213,
acq_224, and acq_280.

But I took another look at the data, expecting to find something else I
might have missed only to realize that it was in fact no differently
formatted than the rest of the data.  The actual culprit is some entries
in the error log at the very end of the file that happen to match the
same formatting rules that the parser is using to process the real log
time series.  This is what I found there, and the boldfaced lines are
really what caused IO-445 in the first place:

ExperimentErrorLog:
	setting cons templatingReagent_0 to ignored 1
	flow 42 flowrate 22.13 too low expected atleast 23.00
	flow 49 flowrate 21.30 too low expected atleast 23.00
	flow 111 flowrate 19.80 too low expected atleast 23.00
	flow 179 flowrate 21.63 too low expected atleast 23.00
	acq_0213.dat: flowTime=77.36 time=04/15/2021 05:04:41
	acq_0224.dat: flowTime=65.57 time=04/15/2021 05:09:24
	acq_0280.dat: flowTime=83.28 time=04/15/2021 05:24:22
	flow 293 flowrate 21.47 too low expected atleast 23.00
	flow 296 flowrate 20.83 too low expected atleast 23.00
	flow 353 flowrate 19.97 too low expected atleast 23.00
	flow 357 flowrate 20.90 too low expected atleast 23.00
	flow 361 flowrate 19.87 too low expected atleast 23.00
	flow 459 flowrate 22.23 too low expected atleast 23.00
	flow 477 flowrate 20.90 too low expected atleast 23.00

It is too risky to modify the explog parser to compensate for this
artifact because too many things depend on it, so the fallback position
for now is to recognize and remove overwritten data points from the
plotted data set.  A handful of missing points among a few hundred
remaining ones should not skew the information here too terribly.