Removing ALLOWED_HOSTS='' key from development overrides
With ALLOWED_HOSTS='', the Django setting turns out to become a
list with one empty string, which will never match ant host's name.

With the variable unset, default behavior kicks in and we end up
with ALLOWED_HOSTS=(), an emty list, and exactlt what is desired
in this particular context.