Drop unimplementable call to get_serial_no() with OneTouch tags
tags.append(get_serial_no(archive_path)

This is a nonsensical inclusion, since the get_serial_number() routine only knows how to locate serial numbers in explog.txt, and OneTouch uploads do not have any kind of explog.txt file to target with that search. The routine has a try catch block that returns the ensuing exception's message about being unable to find an explog.txt to read from instead of the requested serial number, so that's how the error message has wound up as a tag. Craig agreed with me that we likely never had any user ask expecting this search tag to exist--this is a side effect of a misguided attempt to retrieve a serial number that could not have worked without special handling that does not exist.

Craig was fine with skipping any heroic measures attempting to find a serial number for OT devices elsewhere--the OneTouch support lifespan is sufficiently limited, and this method never returned a serial number in the first place anyhow. Commenting out the errant line in a hot fix is a preferred approach.  This is needed because the explog error message has finally grown in length such that it exceeds the databas column size in production, preventing One Touch uploads from occurring at all.